test/requirements.txt
requests==2.31.0
python-jira==3.5.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
tqdm==4.66.1
python-dotenv==1.0.0
lxml==4.9.3
